Converting d3.phylogram.js to a Vega transform
A data transform, that is. Actually two of them (this may change): one
for the main layout and projection, and another for our special link
paths. Since the path geometry is dependent on the phylogram
orientation, I'll probably resort to pre-baking paths and attaching them
directly to links, to be applied later in the rendering phase.