Improve style switcher. Clean up UI.
I'm starting to clarify the distinction between style-guide constraints
and per-illustration style assertions. (These should be separate, so
that style switching doesn't override user decisions, but we should show
warnings where constraints are being violated.)

We now have a simple style resolver to find the effective value for a
given style property on a given element. This should Do The Right Thing
whether the style is defined locally, or in a parent tree or
illustration, and will apply style-guide constraints where needed.