Add support for Newick input and visible SVG output.
Newick trees (pasted text) are converted to Nexson using the opentree
dev site as a service, then cached for fast updates.

SVG output is now available as a "print" option that will instead show
the current SVG into a textarea so the user can save it to their
filesystem.  (This output has been tested in Inkscape and seems to
render properly.)

This commit also adds options to choose the source field (property) for
node labels, and some auto-switching behavior for imported Newick.