Move detailed style editor to main section.
This encourages consistent use of styles for all trees/ornaments/etc.
in an illustration. (Don't Repeat Yourself.) It's also more compatible
with a generic, full-featured style guide that can be re-used across
illustrations. But, it also forces the user to think more indirectly
about how to apply "special" style to chosen elements (define the style
generally, give it a name?, then apply to n elements, adjust to taste).