Switching to jQuery event binding for 'message'
jQuery allows us to unbind and (re)bind to a namespaced event
(message.TreeIllustrator). This should prevent duplicate event listeners
when reloading a script in the notebook, which has caused problems in
the past. The more traditional `addEventListener` approach doesn't allow
un-binding, or even detection of existing listeners on an object, so
hopefully this will give us better results.