An interesting (but failed) experiment...
This was an attempt to quickly generate "proper" handles and hotspots
on-the-fly, as the user is dragging handles in the viewport. The goal
was to show these features using the latest IllustratedTree properties,
without the overhead of a full illustration update, to keep things
lively even for complex illustrations.

Unfortunately, we're blocked at a few turns:

- I attempted to "re-project" hotspots and handles using the existing
  `hotspotGenerator` and `handleGenerator` functions in the phylogram
  data transform. I was able to expose these to the stylist, but the
  transform "remembers" the last instantiated tree and always projects
  its geometry. This might be addressed by storing/caching data for all
  trees encountered within the transform, or keeping each transform
  instance separate, so each only knows about its own tree.

- I tried using Vega 2 signals to allow quick updates to key tree
  properties in the transform, and using Vega's Runtime API to set these
  signals on demand. But attempts (using the Runtime API) to .render()
  or .update() a partial spec have not worked for me, possibly because
  of
    - the odd construction of my data transforms? esp. in regard to the
      passing of data through .add, .update, etc. arrays
    - the fact that our data is rarely in the expected tuple form?
    - our use of cached data in the pipeline?

All this is worth a fresh look later, but for now I'll try a simpler
approach using the known geometry of the IllustratedTree and its known
extents (bounding box) in the current illustration.