Refactoring to use ordered style rules
This is once again emulating CSS rules (ordered, selector-based) but
necessary to capture the expected behavior of default vs. special-case
styles, in a way that's more flexible than simple containment (local
`style` properties for every object) will support.

That said, we might restore per-object `style` as a form of caching, if
dynamic styling is too slow for large trees.